Refactor authentication imports across frontend components to streamline code organization. Update import paths to utilize hooks for authentication and toast notifications, enhancing maintainability and consistency throughout the application.
