+++ frontend/src/components/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from "react-router-dom";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../hooks/useAuth";
 
 const ProtectedRoute = () => {
   const { isAdmin, isLoggedIn } = useAuth();
