+++ frontend/src/hooks/useApi.ts
@@ -1,5 +1,5 @@
 import { useState, useCallback } from 'react';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from './useAuth';
 
 interface UseApiOptions<T = unknown> {
   onSuccess?: (data: T) => void;
