+++ frontend/src/hooks/useToast.ts
@@ -1,0 +1,18 @@
+import { useState } from "react";
+import type { ToastMessage, ToastType } from "../components/common/Toast";
+
+export const useToast = () => {
+  const [toasts, setToasts] = useState<ToastMessage[]>([]);
+
+  const addToast = (message: string, type: ToastType = "info") => {
+    const id = `${Date.now()}-${Math.random()}`;
+    setToasts((prev) => [...prev, { id, message, type }]);
+  };
+
+  const removeToast = (id: string) => {
+    setToasts((prev) => prev.filter((t) => t.id !== id));
+  };
+
+  return { toasts, addToast, removeToast };
+};
+
